backend/pyproject.toml
[project]
name = "weather-logger-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "authlib>=1.6.1",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "fastapi[testing]>=0.116.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]
